refactor: Move disablerule package public
