+++ linter/disablerule/command.go
