+++ linter/disablerule/commands.go
