+++ linter/disablerule/interpreter.go
